requirements.txt
python-dotenv
pandas
matplotlib
pyyaml
requests
numpy